finished clean up of testing
